Refactor Sensors table into a component
